backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
pytest==6.2.4
pytest-pythonpath==0.7.3
pytest-django==4.4.0